pyproject.toml
[tool.poetry]
name = "spotify-etl-aws"
version = "0.1.0"
description = ""
authors = ["danielmschaves <danielmschaves@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
boto3 = "^1.34.64"
requests = "^2.31.0"
pytest = "^8.1.1"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
flask = "^2.0"
duckdb = "^0.9.2"
dbt-duckdb = "^1.7.3"
apache-airflow = "^2.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
